This is the beginning of the new members event.
This is still work in progress.
It has the logic of what has to happen if the user is new, or to check all users to see seniority, etc...
Next step will be to add the proper events that will trigger this.
My idea is to make it run once/twice a day, to not put too much load on the server (it's not a heavy process, but idk how powerful the server is), but the ideal would be to run it once or twice an hour, but I also don't fully know how discord will go about that, if it'll consider that as spam or malicious activity, I guess we'll see.